Seperate input and output processors
Separate Granite 3.2 input and output processors
from the IO processor for following reasons:

- Re-use input or output processor in another IO processor
- Make the IO processor more extensible where input or
output processor can change with configuration
- Export input and output processors like IO processor to be used
individually
