+++ src/granite_io/io/__init__.py
@@ -7,5 +7,7 @@ processing.
 
 # Local
 from granite_io.io.base import InputOutputProcessor  # noqa: F401
+from granite_io.io.base import InputProcessor  # noqa: F401
+from granite_io.io.base import OutputProcessor  # noqa: F401
 from granite_io.io.registry import io_processor, make_io_processor  # noqa: F401
 import granite_io.io.granite_3_2.granite_3_2  # noqa: F401
