+++ src/granite_io/io/consts.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""
+Common shared constants
+"""
+
+# Strings that a Granite 3.2 model uses to separate Chain of Thought output from
+# responses.
+_GRANITE_3_2_COT_START = "Here is my thought process:"
+_GRANITE_3_2_COT_END = "Here is my response:"
