+++ tests/io/output_processors/granite_3_2/testdata/test_output_with_invalid_citation.txt
